refactor: move copyInviteLink to server invite utils file
