+++ src/components/ServerInvites/ServerInviteRow.tsx
@@ -1,41 +1,22 @@
-import { ApolloCache } from "@apollo/client";
 import { Assignment } from "@mui/icons-material";
 import { Box, MenuItem, styled, TableRow } from "@mui/material";
-import produce from "immer";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { toastVar } from "../../apollo/cache";
 import {
   ServerInviteCardFragment,
-  ServerInvitesDocument,
   ServerInvitesQuery,
   useDeleteServerInviteMutation,
 } from "../../apollo/gen";
-import { TypeNames } from "../../constants/common.constants";
 import { ServerPermissions } from "../../constants/role.constants";
+import { copyInviteLink } from "../../utils/server-invite.utils";
 import { timeFromNow } from "../../utils/time.utils";
 import { getUserProfilePath } from "../../utils/user.utils";
 import ItemMenu from "../Shared/ItemMenu";
 import Link from "../Shared/Link";
 import SharedTableCell from "../Shared/TableCell";
 import UserAvatar from "../Users/UserAvatar";
-
-export const removeServerInvite = (id: number) => (cache: ApolloCache<any>) => {
-  cache.updateQuery<ServerInvitesQuery>(
-    { query: ServerInvitesDocument },
-    (invitesData) =>
-      produce(invitesData, (draft) => {
-        if (!draft) {
-          return;
-        }
-        const index = draft.serverInvites.findIndex((p) => p.id === id);
-        draft.serverInvites.splice(index, 1);
-      })
-  );
-  const cacheId = cache.identify({ id, __typename: TypeNames.ServerInvite });
-  cache.evict({ id: cacheId });
-  cache.gc();
-};
+import { removeServerInvite } from "./ServerInviteCard";
 
 const TableCell = styled(SharedTableCell)(({ theme }) => ({
   color: theme.palette.text.secondary,
@@ -77,9 +58,7 @@ const ServerInviteRow = ({
     });
 
   const handleCopyLink = async () => {
-    const inviteLink = `${window.location.origin}/i/${token}`;
-    await navigator.clipboard.writeText(inviteLink);
-
+    await copyInviteLink(token);
     toastVar({
       title: t("invites.prompts.copiedToClipboard"),
       status: "success",
