+++ src/utils/server-invite.utils.ts
@@ -1,0 +1,4 @@
+export const copyInviteLink = async (token: string) => {
+  const inviteLink = `${window.location.origin}/i/${token}`;
+  await navigator.clipboard.writeText(inviteLink);
+};
